Moved TargetHoverfly into target.go
