+++ hoverctl/wrapper/config.go
@@ -9,14 +9,6 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-type TargetHoverfly struct {
-	Name      string
-	Host      string `yaml:"host"`
-	AdminPort int    `yaml:"admin.port"`
-	AuthToken string `yaml:"auth.token"`
-	Pid       int    `yaml:"pid"`
-}
-
 type Flags []string
 
 type Config struct {
